+++ demo/FormDemo.jsx
@@ -9,8 +9,8 @@
 
 /* eslint-disable react/jsx-no-bind */
 
-import classnames from 'classnames';
 import Button from 'uxcore-button';
+import Icon from 'uxcore-icon';
 import React from 'react';
 import Form from '../src';
 
@@ -39,7 +39,7 @@ const {
 const CheckboxItem = CheckboxGroupFormField.Item;
 const RadioItem = RadioGroupFormField.Item;
 const PickItem = PickableFormField.Item;
-const { Count, LeftAddon, RightAddon } = InputFormField;
+const { Count, RightAddon } = InputFormField;
 const { TextAreaCount } = TextAreaFormField;
 const Option = SelectFormField.Option;
 const CustomField = createFormField();
@@ -78,6 +78,8 @@ class Demo extends React.Component {
         xj: '西京',
       },
       mode: Constants.MODE.EDIT,
+      size: 'middle',
+      verticalAlign: true,
     };
   }
 
@@ -264,9 +266,9 @@ class Demo extends React.Component {
       <div className="demo">
         <Form
           ref="form"
-          size="large"
+          size={me.state.size}
           instantValidate
-          verticalAlign={false}
+          verticalAlign={me.state.verticalAlign}
           jsxmode={me.state.mode}
           jsxvalues={me.state.jsxvalues}
           jsxonChange={me.handleChange.bind(me)}
@@ -321,7 +323,9 @@ class Demo extends React.Component {
               >
                 提交
               </Button>
-              <Button size="medium" type="secondary" action="reset">取消</Button>
+              <Button size="medium" type="secondary" action="reset">
+                取消
+              </Button>
             </ButtonGroupFormField>
           </FormRow>
           <FormRow>
@@ -332,22 +336,21 @@ class Demo extends React.Component {
             </RadioGroupFormField>
             <CheckboxGroupFormField jsxname="checkbox" jsxlabel="复选框">
               <CheckboxItem
-                value="air" text="天空"
+                value="air"
+                text="天空"
               />
               <CheckboxItem value="sea" text="大海" />
             </CheckboxGroupFormField>
           </FormRow>
           <InputFormField inputBoxMaxWidth="middle" jsxname="pass" jsxlabel="请输入密码" inputType={me.state.showPass ? 'text' : 'password'}>
-            <LeftAddon>
-              <i
-                className={classnames({
-                  'kuma-icon': true,
-                  'kuma-icon-privacy': !me.state.showPass,
-                  'kuma-icon-public': me.state.showPass,
-                })}
+            <RightAddon>
+              <Icon
+                usei
+                style={{ cursor: 'pointer' }}
+                name={!me.state.showPass ? 'xianshi' : 'yincang'}
                 onClick={me.handleShowPassChange.bind(me)}
               />
-            </LeftAddon>
+            </RightAddon>
           </InputFormField>
           <FormRow>
             <SwitchFormField
@@ -356,16 +359,18 @@ class Demo extends React.Component {
               checkedChildren="显示"
               unCheckedChildren="隐藏"
             />
-            {/* <PickableFormField
+            <PickableFormField
               jsxlabel="筛选"
               jsxname="pick"
               multiple
               type="hook"
             >
               {itemsData.map((item, index) => (
-                <PickItem key={index} value={item.value}>{item.text}</PickItem>
+                <PickItem key={index} value={item.value}>
+                  {item.text}
+                </PickItem>
               ))}
-            </PickableFormField> */}
+            </PickableFormField>
           </FormRow>
           <TextAreaFormField
             jsxname="textArea"
@@ -436,9 +441,15 @@ class Demo extends React.Component {
             jsxname="option"
             jsxlabel="传 option"
           >
-            <Option value="1">第一个选项</Option>
-            <Option value="2">第二个选项</Option>
-            <Option value="3">第三个选项</Option>
+            <Option value="1">
+              第一个选项
+            </Option>
+            <Option value="2">
+              第二个选项
+            </Option>
+            <Option value="3">
+              第三个选项
+            </Option>
           </SelectFormField>
           <FormRowTitle jsxtitle="级联类" />
           <DateFormField
@@ -452,7 +463,7 @@ class Demo extends React.Component {
               validator: (value) => {
                 if (typeof value === 'undefined') {
                   return false;
-                } else if (Array.isArray(value) && value.length < 2) {
+                } if (Array.isArray(value) && value.length < 2) {
                   return false;
                 }
                 for (let i = 0, l = value.length; i < l; i++) {
@@ -487,16 +498,29 @@ class Demo extends React.Component {
           <CustomField jsxname="custom" jsxlabel="定制" placeholder="test" />
 
           <ButtonGroupFormField>
-            <Button size="medium" onClick={me.handleClick.bind(me)}>提交</Button>
-            <Button size="medium" type="secondary" action="reset">取消</Button>
-            <Button type="secondary" onClick={me.handleSetValues.bind(me)}>手动setValues</Button>
+            <Button size="medium" onClick={me.handleClick.bind(me)}>
+              提交
+            </Button>
+            <Button size="medium" type="secondary" action="reset">
+              取消
+            </Button>
+            <Button type="secondary" onClick={me.handleSetValues.bind(me)}>
+              手动setValues
+            </Button>
             <Button
               size="medium"
               type="secondary"
               onClick={me.handleValueChange.bind(me)}
-            >修改 props</Button>
-            <Button type="secondary" onClick={me.changeMode.bind(me)}>转变模式</Button>
-            <Button type="secondary" onClick={me.update.bind(me)}>强制刷新</Button>
+            >
+              修改 props
+
+            </Button>
+            <Button type="secondary" onClick={me.changeMode.bind(me)}>
+              转变模式
+            </Button>
+            <Button type="secondary" onClick={me.update.bind(me)}>
+              强制刷新
+            </Button>
           </ButtonGroupFormField>
         </Form>
       </div>