+++ src/FormRowTitle.jsx
@@ -8,31 +8,37 @@ import classnames from 'classnames';
 import Title from 'uxcore-title';
 import PropTypes from 'prop-types';
 
-const FormRowTitle = (props) => {
-  const length = React.Children.count(props.children);
+const FormRowTitle = ({
+  children, className, prefixCls, jsxtitle,
+}) => {
+  const length = React.Children.count(children);
   return (
     <div
       className={classnames({
-        [props.className]: !!props.className,
-        [props.jsxprefixCls]: true,
+        [className]: !!className,
+        [prefixCls]: true,
       })}
     >
-      <Title type="secondary">{props.jsxtitle}</Title>
-      {!!length && React.Children.map(props.children, item => item)}
+      <Title type="secondary">
+        {jsxtitle}
+      </Title>
+      {!!length && React.Children.map(children, item => item)}
     </div>
   );
 };
 
 FormRowTitle.propTypes = {
-  jsxprefixCls: PropTypes.string,
+  prefixCls: PropTypes.string,
   jsxtitle: PropTypes.string,
   className: PropTypes.string,
   children: PropTypes.any,
 };
 
 FormRowTitle.defaultProps = {
-  jsxprefixCls: 'kuma-uxform-row-title',
+  prefixCls: 'kuma-uxform-row-title',
   jsxtitle: '',
+  children: undefined,
+  className: '',
 };
 
 FormRowTitle.displayName = 'FormRowTitle';