+++ src/KeyCode.js
@@ -1,6 +1,3 @@
 export default {
-    /*
-     * Enter
-     */
   Enter: 13,
 };